requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
diskcache==5.6.3
disnake==2.9.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
haystack-ai==2.6.1
haystack-experimental==0.2.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-imports==0.3.1
llama-cpp-haystack==0.4.1
llama_cpp_python==0.3.1
lxml==5.3.0
MarkupSafe==3.0.1
monotonic==1.6
more-itertools==10.5.0
multidict==6.1.0
networkx==3.4.1
numpy==1.26.4
openai==1.51.2
pandas==2.2.3
pdfminer.six==20240706
posthog==3.7.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.0
